Update to reflect review feedback.
There are just three changes here:
- In chplsys.c, only define getPUsPerCore() for Linux targets, not for
  all non-Apple targets.  It's only called on Linux targets.
- Also in chplsys.c, collapse the multi-space runs into single spaces in
  the sscanf() format strings used to read values from /proc/cpuinfo.  A
  single space in the format is sufficient to match any amount of white
  space in the input, and using multi-space runs implies more input
  matching precision than is actually happening.
- Explicitly cast return values where they don't match the function
  return type, in both chplsys.c and the tasking layers.

+++ runtime/src/tasks/massivethreads/tasks-massivethreads.c
@@ -483,7 +483,7 @@ uint32_t chpl_task_getMaxPar(void) {
   // from gaining by using them.  So, we just return the number of
   // cores.
   //
-  return chpl_getNumCoresOnThisNode();
+  return (uint32_t) chpl_getNumCoresOnThisNode();
 }
 
 c_sublocid_t chpl_task_getNumSublocales(void)
